kgo: fix KIP-890 handling of the transaction.version feature
Kafka introduced "features" in the ApiVersions response a long time ago
during Kafka 2.6. This was never used in clients, and I did not add
support for features to franz-go.

Well, Kafka 4.0 finally started using features in clients.

KIP-890 -- which does not mention the exact feature name
"transaction.version" -- requires clients to only opt into "part 2" of
the KIP if transaction.version >= 2.

What this means is that if you were talking to Kafka 4 but had not yet
enabled transactions in the cluster, transactions would not work.

I audited a bit of the Java client code and discovered a few other
things that I hadn't implemented -- for these, I point to the lack of
actually calling out protocol version bumps in the KIP and mentioning
the changes in like, 5 words:

* AddOffsetsToTxn is no longer sent if we use TxnOffsetCommit v5+ (same
  principle as AddPartitionsToTxn not being sent for Produce v12+)

* TxnOffsetCommit is pinned to v4 unless we detect kip890p2

* EndTxn is pinned to v4 until we BEGIN a transaction with kip890p2

* If we end and were pinned to v4 but then detect the cluster supports
  kip890p2, we re-init our producer ID (Java client does this to
  simplify things as we opt into p2)

* Produce is pinned to v11 if we are transactional and kip890p2 is not
  supported (same as Java client)

Closes #1017.