kgo: remove pinReq wrapping request
Every time I've added some behavior that uses pinReq in a new way,
there's been a bug. The fundamental problem is that pinReq wrapping a
kmsg.Request: removes any extra optional interface (is txn req, is group
req); breaks all type switches; is confusing to handle.

This moves pinReq from a wrapper to a value to carry along with the
context. This makes it a lot easier to reason about how pinReq will
break a reqest, because the answer is: it wont.

This requires a big switchup in how versions are handled in broker.go;
IMO the code actually is easier to reason about with min&max.

Locally tested against Kafka 4, Kafka 3.9, and Kafka 2.8; integ tests
will run against 4 and 3.9. I also inserted some panics to triple verify
that pinReq in sharding was actually used somewhere -- it is, for
AddPartitionsToTxn, where we pinMax on pre-4.0 versions.

Closes #1017.