kadm: bump deps, new APIs to support better lag
If a group is actively rebalancing or cooperative but rebalancing for
the first time, the group will not have assigned partitions. We can look
at what each member WANTS to consume and use that as further input into
lag calculations.

This adds a few helper APIs to make further lag calculation easy
(JoinTopics alongside AssignedPartitions, and a MergeTopics API on
TopicsSet).

When iterating over all members, after we iterate over their assigned
partitions, we can iterate over their interested topics and prime the
lag map. If the topic is not in the lag map, add it. The existing logic
at the end of the function goes over all topics in the lag map and
calcuates lag per partition for all missing partitions.