kgo: bugfix batched FindCoordinator requests against older brokers
All credit to @douglasbouttell for the excellent find here.

Batched FindCoordinator requests against brokers that do not support
batched requests no longer worked at of 1.19.3. The problem was the
removal of pinReq, and well, a bug in the full removal.

This fixes the problem and adds a test test that fails without the fix.

I checked all linters I know of (opting into all revive, staticcheck,
gocritic, golangci-lint linters) and none of them catch this, for valid
reasons. It's hard to check.

I then went back and forth with Claude to generate some code that would
check it, and surprisingly came up with something. However, since it
generates a _few_ false positives and I don't want to learn enough of
the ast code to fully maintain the linter, I'm going to comment the code
on #1034 and this PR and if anybody wants to take it and run with it,
feel free. The one other variable the lint flagged was a false positive.

Closes #1034.