kgo: add ProduceSyncBatch
Local benchmarking shows this to be 50% faster in high tput use cases,
but it's not without downsides.

The implementation is a bit tricky to wire in; in reality, this would be
easier to implement if the existing code was better. As is, too many
functions have too many side effects, which gets a bit difficult to
reason about. It'll be good to pull out the side effects over time as
more work goes into these areas. The original implementation was
basically written top down to do the thing (producing), and then a LOT
of behavior got added in over time.

Anyway, what remains is adding more tests for this. I did originally
implement this as ProduceSync (replacing the original produce sync
implementation) and then fixed code so that all tests *using* produce
sync worked against this. But, I need more dedicated tests.

Relates to #1064, #1074.