kgo broker: also add eof error checking to api version negotiating
This also backs out of the last change to the logging in handling
responses. TLS errors will happen during api version negotiation, so if
we encounter EOF later, it either HAS to be SASL, or we're talking to
such an old broker that it doesn't support API version negotiation.
We'll assume SASL.