kfake: fix new test, bump franz-go to latest
The new test originally used BrokerMetadata which returns broker
information only (no topics).

Checking for all topics was added as a suggestion, but BrokerMetadata
was kept -- this was masked because I requested moving the kfake test to
a new PR, and then merged it pre-emptively because eh, it's kinda a
pain to have kfake be dependent on franz-go and then adding tests for a
new franz-go release that hasn't been cut yet.

Anyway, in the process of bumping the dep, the test started failing;
this is the small fix.