kgo: fix data race when the linger timer fires
If a partition is moving from one broker to another via a metadata
update AT THE SAME TIME a linger timer is firing, there is a data race:
The recBuf sink pointer is changing at the same time it is being read.

The fix is to use the existing unlingerAndManuallyDrain, which locks on
draining and clears the lingering pointer (which was missing before).