kgo source: expired old preferred replicas while creating req, not handling resp
We periodically re-fetch from the leader for preferred replicas, just in
case the leader makes a decision that a new different replica is
preferable.

As previously implemented, we processed moving back to the leader at the
end of handling a fetch response. The implementation was bad: we'd move
back to the leader but use the offset we just fetched, while also still
still returning the data we just fetched from the follower.

We move the logic now to creating requests. This makes the logic
actually accurate -- we move the partition to the leader with the offset
we were about to fetch (and now the leader, and redirection back to the
follower, all has the correct offset) -- and much easier to reason
about.

This also modifies the kfake test for this behavior to actually check
the offsets that are being consumed. The test fails before the patch,
and succeeds after the patch.