kgo: retry dial errors on sharded requests if possible
Specifically, if a sharded request has a portion that can go to any
broker, and the chosen broker fails due to a dial error, we currently
hard fail the request (since dial errors are not retryable). However,
dial errors are retryable on other brokers. We can retry the sharded
request and we try (up to 3x) to receive a different broker.

We could go deeper with this and make sure we *never* retry a broker
that we previously deemed skippable -- i.e., make sure we don't just
skip all the brokers repeatedly in a loop. That can be a follow up task
if needed (and, at that point, would probably be worth wiring into the
existing use of shouldRetryNext on the `retryable` type).

I'm a little worried about just loop skipping around the entire cluster,
but really that means you need to have dial failures for every broker
address, which at that point is likely due to invalid seed addresses
and/or DNS issues. Both are ideally caught quickly enough.