kgo: support arbitrary Kafka RPCs
Sometimes it is desireable to have RPCs defined outside the set that
franz-go knows about. This patch enables that by storing API versions
in a map instead of a fixed length array.

Alternative solutions that achieve this same goal I would be open to
implementing instead (since one has to use a lot of low level APIs for
all this anyways):
* Just allow skipping API version checks for keys outside the range kgo
  supports via some configuration option
* Only skip the "brokerMin/brokerMax" checks if the issued request is
  outside the range kgo knows about. At this point it's clear the user
  is issuing a custom command, and it's sort of on them to check the
  versions...
Let me know if either option is preferred, then I'd happily update this
patch.