kgo: fix data race with replica fetching / moving; kfake: add test
When moving a cursor to a new preferred replica, we need to wait to add
the cursor to the new source until AFTER we are completely done
modifying or accessing fields on the cursor.

A few fields were modified after adding (race), and the defer
allowUsable accessed the source, which, once on the new source, could be
concurrently modified (if moving again, pathologically).

Thanks to @rpadaki for the bug report & key finding pointing in the
right direction.

Closes #1167.