Revert "kgo: avoid updating metadata before issuing ListOffsets / OffsetForLeaderEpoch"
Meant to merge #1174.

Upon further analysis, the metadata update is called anyway: sharding
uses fetchMappedMetadata, which issues a metadata request if it's been
more than metadataMinAge. The downside of doing this only through
fetchMappedMetadata is that that codepath *only* updates the map, not
the client's internal metadata that it uses for producing and consuming.
So, not only do we not reduce the number of metadata requests, we
actually propagate the information less with the commit.

This reverts commit 88696a4aae36746d9d20e6c2c6850ce250fa6f3a.